+++ lib/constant/assets.dart
@@ -1,4 +1,4 @@
-class assets {
+class Assets {
   //images
   static const imageDir = 'assets/images';
   static const imageWelcome = '$imageDir/welcome.png';
