+++ pubspec.yaml
@@ -60,28 +60,26 @@ flutter:
   assets:
     - assets/images/
 
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware
+  fonts:
+    - family: Prompt
+      fonts:
+        - asset: assets/fonts/Prompt/Prompt-Black.ttf
+        - asset: assets/fonts/Prompt/Prompt-BlackItalic.ttf
+        - asset: assets/fonts/Prompt/Prompt-Bold.ttf
+        - asset: assets/fonts/Prompt/Prompt-BoldItalic.ttf
+        - asset: assets/fonts/Prompt/Prompt-ExtraBold.ttf
+        - asset: assets/fonts/Prompt/Prompt-ExtraBoldItalic.ttf
+        - asset: assets/fonts/Prompt/Prompt-ExtraLight.ttf
+        - asset: assets/fonts/Prompt/Prompt-ExtraLightItalic.ttf
+        - asset: assets/fonts/Prompt/Prompt-Italic.ttf
+        - asset: assets/fonts/Prompt/Prompt-Light.ttf
+        - asset: assets/fonts/Prompt/Prompt-LightItalic.ttf
+        - asset: assets/fonts/Prompt/Prompt-Medium.ttf
+        - asset: assets/fonts/Prompt/Prompt-Regular.ttf
+        - asset: assets/fonts/Prompt/Prompt-SemiBold.ttf
+        - asset: assets/fonts/Prompt/Prompt-SemiBoldItalic.ttf
+        - asset: assets/fonts/Prompt/Prompt-Thin.ttf
+        - asset: assets/fonts/Prompt/Prompt-ThinItalic.ttf
 
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
-
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
